backend/pyproject.toml
[tool.poetry]
name = "shopmood"
version = "0.1.0"
description = "An app that suggests products based on user's mood using emotion recognition technology."
authors = ["AJB"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.1"
Pillow = "^8.3.2"
pydantic = "^1.8.2"
bottleneck = "^1.3.2"
tensorflow = "^2.4.0"
keras = "^2.4.3"
numpy = "^1.21.2"
matplotlib = "^3.4.3"
opencv-python = "^4.5.3"
opencv-contrib-python = "^4.5.3"
scipy = "^1.7.1"
scikit-learn = "^0.24.2"
scikit-image = "^0.18.3"
tqdm = "^4.62.3"
tensorboard = "^2.6.0"
libclang = "15.0.6.1"
jupyterlab = "^4.0.3"
openai = "^0.27.8"
transformers = "^4.31.0"
torch = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
